+++ raphtory/src/vectors/vectorised_graph.rs
@@ -1,228 +1,21 @@
 use crate::{
-    // core::entities::nodes::node_ref::AsNodeRef,
     core::{entities::nodes::node_ref::AsNodeRef, utils::errors::GraphResult},
-    db::{
-        api::{
-            storage::graph::edges::edge_storage_ops::EdgeStorageOps,
-            view::{DynamicGraph, IntoDynamic, StaticGraphViewOps},
-        },
-        graph::{edge::EdgeView, node::NodeView},
-    },
+    db::api::view::{DynamicGraph, IntoDynamic, StaticGraphViewOps},
     prelude::*,
     vectors::{
-        embedding_cache::EmbeddingCache, template::DocumentTemplate, utils::find_top_k, Embedding,
-        EmbeddingFunction,
+        embedding_cache::EmbeddingCache, embeddings::compute_embeddings,
+        template::DocumentTemplate, utils::find_top_k, Embedding,
     },
 };
-use arroy::{distances::Cosine, Database as ArroyDatabase, Reader, Writer};
 use futures_util::StreamExt;
-use rand::{rngs::StdRng, SeedableRng};
-use std::{
-    collections::HashSet,
-    path::PathBuf,
-    sync::{Arc, OnceLock},
-};
-use tempfile::TempDir;
+use std::{path::PathBuf, sync::Arc};
 
 use super::{
-    utils::apply_window, vector_selection::VectorSelection, vectorisable::compute_embeddings,
+    db::{EdgeDb, NodeDb},
+    embeddings::EmbeddingFunction,
+    utils::apply_window,
+    vector_selection::VectorSelection,
 };
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
-pub(crate) enum EntityRef {
-    Node(usize), // TODO: store this as a u32
-    Edge(usize),
-}
-
-impl EntityRef {
-    // TODO: there are some places where I should be using this and Im not
-    pub(crate) fn as_node_view<G: StaticGraphViewOps>(&self, view: &G) -> Option<NodeView<G>> {
-        let node = self.as_node(view)?;
-        view.node(node)
-    }
-
-    // TODO: there are some places where I should be using this and Im not
-    pub(crate) fn as_edge_view<G: StaticGraphViewOps>(&self, view: &G) -> Option<EdgeView<G>> {
-        let (src, dst) = self.as_edge(view)?;
-        view.edge(src, dst)
-    }
-
-    pub(crate) fn as_node<G: StaticGraphViewOps>(&self, view: &G) -> Option<GID> {
-        if let EntityRef::Node(id) = self {
-            Some(view.node_id(raphtory_api::core::entities::VID(*id)))
-        } else {
-            None
-        }
-    }
-
-    pub(crate) fn as_edge<G: StaticGraphViewOps>(&self, view: &G) -> Option<(GID, GID)> {
-        if let EntityRef::Edge(id) = self {
-            let edge = view.core_edge((*id).into());
-            let src = view.node_id(edge.src());
-            let dst = view.node_id(edge.dst());
-            Some((src, dst))
-        } else {
-            None
-        }
-    }
-
-    // fn as_usize(&self) -> usize {
-    //     match self {
-    //         EntityRef::Node(id) => *id,
-    //         EntityRef::Edge(id) => *id,
-    //     }
-    // }
-
-    fn as_u32(&self) -> u32 {
-        match self {
-            EntityRef::Node(id) => *id as u32,
-            EntityRef::Edge(id) => *id as u32,
-        }
-    }
-}
-
-#[derive(Clone)]
-pub(super) struct NodeDb(pub(super) VectorDb);
-
-impl VectorSearch for NodeDb {
-    fn get_db(&self) -> &VectorDb {
-        &self.0
-    }
-
-    fn into_entity_ref(id: u32) -> EntityRef {
-        EntityRef::Node(id as usize)
-    }
-
-    fn view_has_entity<G: StaticGraphViewOps>(entity: &EntityRef, view: &G) -> bool {
-        view.has_node(entity.as_node(view).unwrap())
-    }
-
-    fn all_valid_entities<G: StaticGraphViewOps>(view: G) -> impl Iterator<Item = usize> {
-        view.nodes().into_iter().map(|node| node.node.index())
-    }
-}
-
-#[derive(Clone)]
-pub(super) struct EdgeDb(pub(super) VectorDb);
-
-impl VectorSearch for EdgeDb {
-    fn get_db(&self) -> &VectorDb {
-        &self.0
-    }
-
-    fn into_entity_ref(id: u32) -> EntityRef {
-        EntityRef::Edge(id as usize)
-    }
-
-    fn view_has_entity<G: StaticGraphViewOps>(entity: &EntityRef, view: &G) -> bool {
-        let (src, dst) = entity.as_edge(view).unwrap(); // TODO: remove this?
-        view.has_edge(src, dst) // FIXME: there should be a quicker way!!!!!!!!!!!!!!!!!!!
-    }
-
-    fn all_valid_entities<G: StaticGraphViewOps>(view: G) -> impl Iterator<Item = usize> {
-        view.edges().into_iter().map(|edge| edge.edge.pid().0)
-    }
-}
-
-// FIXME: remove unwraps in here
-// TODO: rename this to GraphVectorSearch
-// TODO: merge this and VectorDb !!!!!!!!!!!!!!!!!???????
-pub(super) trait VectorSearch {
-    fn get_db(&self) -> &VectorDb;
-    fn into_entity_ref(id: u32) -> EntityRef;
-    fn view_has_entity<G: StaticGraphViewOps>(entity: &EntityRef, view: &G) -> bool;
-    fn all_valid_entities<G: StaticGraphViewOps>(view: G) -> impl Iterator<Item = usize> + 'static;
-
-    fn top_k<G: StaticGraphViewOps>(
-        &self,
-        query: &Embedding,
-        k: usize,
-        view: Option<G>,
-        filter: Option<HashSet<EntityRef>>,
-    ) -> impl Iterator<Item = (EntityRef, f32)> {
-        let candidates: Option<Box<dyn Iterator<Item = u32>>> = match (view, filter) {
-            (None, None) => None,
-            (view, Some(filter)) => Some(Box::new(
-                filter
-                    .into_iter()
-                    .filter(move |entity| {
-                        view.as_ref()
-                            .map_or(true, |view| Self::view_has_entity(entity, view))
-                    })
-                    .map(|entity| entity.as_u32()),
-            )),
-            (view, None) => Some(Box::new(
-                Self::all_valid_entities(view.unwrap()).map(|id| id as u32), // FIXME: unwrap here doesnt make sense
-            )),
-        };
-        self.top_k_with_candidates(query, k, candidates)
-    }
-
-    fn top_k_with_candidates(
-        &self,
-        query: &Embedding,
-        k: usize,
-        candidates: Option<impl Iterator<Item = u32>>,
-    ) -> impl Iterator<Item = (EntityRef, f32)> {
-        let db = self.get_db();
-        let rtxn = db.env.read_txn().unwrap();
-        // FIXME: if the db has no edges, I get a MissingMetadata here. Handle that properly,
-        // Maybe the edge db should not exist at all if there are no edge embeddings
-        // because there might be some errors other than MissingMetadata
-        let vectors = if let Ok(reader) = Reader::open(&rtxn, 0, db.vectors) {
-            let mut query_builder = reader.nns(k);
-            let candidates = candidates
-                .map(|filter| roaring::RoaringBitmap::from_iter(filter.map(|id| id as u32)));
-            let query_builder = if let Some(filter) = &candidates {
-                query_builder.candidates(filter)
-            } else {
-                &query_builder
-            };
-            query_builder.by_vector(&rtxn, query.as_ref()).unwrap()
-        } else {
-            vec![]
-        };
-        vectors
-            .into_iter()
-            // for arroy, distance = (1.0 - score) / 2.0, where score is cosine: [-1, 1]
-            .map(|(id, distance)| (Self::into_entity_ref(id), 1.0 - 2.0 * distance))
-    }
-}
-
-#[derive(Clone)]
-pub(crate) struct VectorDb {
-    // FIXME: save index value in here !!!!!!!!!!!!!!!!!!!
-    pub(crate) vectors: ArroyDatabase<Cosine>, // TODO: review is this safe to clone? does it point to the same thing?
-    pub(crate) env: heed::Env,
-    pub(crate) _tempdir: Option<Arc<TempDir>>, // do I really need, is the file open not enough
-    pub(crate) dimensions: OnceLock<usize>,
-}
-
-// TODO: merge this with the above
-impl VectorDb {
-    pub(super) fn insert_vector(&self, id: usize, embedding: &Embedding) {
-        // FIXME: remove unwraps
-        let mut wtxn = self.env.write_txn().unwrap();
-
-        let dimensions = self.dimensions.get_or_init(|| embedding.len());
-        let writer = Writer::<Cosine>::new(self.vectors, 0, *dimensions);
-        writer
-            .add_item(&mut wtxn, id as u32, embedding.as_ref())
-            .unwrap();
-
-        let mut rng = StdRng::from_entropy();
-        writer.builder(&mut rng).build(&mut wtxn).unwrap();
-
-        wtxn.commit().unwrap();
-    }
-
-    pub(super) fn get_id(&self, id: usize) -> Embedding {
-        let rtxn = self.env.read_txn().unwrap();
-        let reader = Reader::open(&rtxn, 0, self.vectors).unwrap();
-        let vector = reader.item_vector(&rtxn, id as u32).unwrap().unwrap();
-        vector.into()
-    }
-}
 
 #[derive(Clone)]
 pub struct VectorisedGraph<G: StaticGraphViewOps> {
